requirements.txt
Flask~=3.1.1
Flask-Caching~=2.3.1
python-memcached~=1.62
httpx
pyyaml~=6.0.2
gunicorn~=23.0.0
